move docker HTTP client code into container_runtimes folder (#289)
