+++ container_runtimes/docker/http/asserts/dockerfiles/base/d/Dockerfile
