+++ container_runtimes/docker/http/asserts/dockerfiles/base/go/Dockerfile
