+++ container_runtimes/docker/http/asserts/dockerfiles/base/java/Dockerfile
