+++ container_runtimes/docker/http/asserts/dockerfiles/base/node/Dockerfile
