+++ container_runtimes/docker/http/asserts/dockerfiles/base/python/Dockerfile
