+++ container_runtimes/docker/http/asserts/dockerfiles/base/rust/config
