+++ container_runtimes/docker/http/stop_test.go
