+++ deploy/docker/docker.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	dockerTypes "github.com/docker/docker/api/types"
-	runtime "github.com/metrue/fx/container_runtimes/docker"
+	runtime "github.com/metrue/fx/container_runtimes/docker/sdk"
 	"github.com/metrue/fx/deploy"
 	"github.com/metrue/fx/utils"
 )
