+++ handlers/deploy.go
@@ -7,9 +7,9 @@ import (
 	"os"
 
 	"github.com/apex/log"
-	"github.com/metrue/fx/api"
 	"github.com/metrue/fx/config"
 	"github.com/metrue/fx/constants"
+	api "github.com/metrue/fx/container_runtimes/docker/http"
 	"github.com/metrue/fx/deploy"
 	dockerDeployer "github.com/metrue/fx/deploy/docker"
 	k8sDeployer "github.com/metrue/fx/deploy/kubernetes"
