+++ handlers/down.go
@@ -2,9 +2,9 @@ package handlers
 
 import (
 	"github.com/apex/log"
-	"github.com/metrue/fx/api"
 	"github.com/metrue/fx/config"
 	"github.com/metrue/fx/constants"
+	api "github.com/metrue/fx/container_runtimes/docker/http"
 	"github.com/pkg/errors"
 	"github.com/urfave/cli"
 )
