+++ handlers/image.go
@@ -7,9 +7,9 @@ import (
 
 	"github.com/apex/log"
 	"github.com/google/uuid"
-	"github.com/metrue/fx/api"
 	"github.com/metrue/fx/config"
 	"github.com/metrue/fx/constants"
+	api "github.com/metrue/fx/container_runtimes/docker/http"
 	"github.com/metrue/fx/packer"
 	"github.com/metrue/fx/provision"
 	"github.com/metrue/fx/utils"
