+++ handlers/up.go
@@ -5,9 +5,9 @@ import (
 	"io/ioutil"
 
 	"github.com/apex/log"
-	"github.com/metrue/fx/api"
 	"github.com/metrue/fx/config"
 	"github.com/metrue/fx/constants"
+	api "github.com/metrue/fx/container_runtimes/docker/http"
 	"github.com/metrue/fx/packer"
 	"github.com/metrue/fx/provision"
 	"github.com/metrue/fx/types"
